refactor: Answer 관련 도메인 패키지 이동
