+++ src/main/java/com/samhap/kokomen/answer/domain/AnswerRank.java
@@ -1,4 +1,4 @@
-package com.samhap.kokomen.interview.domain;
+package com.samhap.kokomen.answer.domain;
 
 import lombok.Getter;
