+++ src/main/java/com/samhap/kokomen/interview/domain/InterviewMessagesFactory.java
@@ -1,5 +1,6 @@
 package com.samhap.kokomen.interview.domain;
 
+import com.samhap.kokomen.answer.domain.Answer;
 import com.samhap.kokomen.interview.external.dto.request.GptMessage;
 import java.util.ArrayList;
 import java.util.List;
