+++ src/test/java/com/samhap/kokomen/global/fixture/interview/AnswerFixtureBuilder.java
@@ -1,7 +1,7 @@
 package com.samhap.kokomen.global.fixture.interview;
 
-import com.samhap.kokomen.interview.domain.Answer;
-import com.samhap.kokomen.interview.domain.AnswerRank;
+import com.samhap.kokomen.answer.domain.Answer;
+import com.samhap.kokomen.answer.domain.AnswerRank;
 import com.samhap.kokomen.interview.domain.Question;
 
 public class AnswerFixtureBuilder {
