+++ src/test/java/com/samhap/kokomen/global/fixture/interview/GptResponseFixtureBuilder.java
@@ -1,6 +1,6 @@
 package com.samhap.kokomen.global.fixture.interview;
 
-import com.samhap.kokomen.interview.domain.AnswerRank;
+import com.samhap.kokomen.answer.domain.AnswerRank;
 import com.samhap.kokomen.interview.external.dto.response.Choice;
 import com.samhap.kokomen.interview.external.dto.response.GptFunctionCall;
 import com.samhap.kokomen.interview.external.dto.response.GptResponse;
