+++ src/test/java/com/samhap/kokomen/interview/domain/InterviewMessagesFactoryTest.java
@@ -2,6 +2,7 @@ package com.samhap.kokomen.interview.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.samhap.kokomen.answer.domain.Answer;
 import com.samhap.kokomen.global.fixture.interview.AnswerFixtureBuilder;
 import com.samhap.kokomen.global.fixture.interview.InterviewFixtureBuilder;
 import com.samhap.kokomen.global.fixture.interview.QuestionFixtureBuilder;
